pyproject.toml
[project]
name = "direct-agent"
version = "0.1.0"
description = "FastAPI 백엔드 서버"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.35.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "Pillow>=10.0.0",
    "urllib3>=2.0.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.10.1",
    "google-genai>=1.35.0",
    "boto3>=1.40.27",
    "google-api-python-client>=2.181.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
]

[project.scripts]
start = "uvicorn src.main:app --reload --host 0.0.0.0 --port 8000"
